+++ src/main/java/kr/co/tododeungjang/web/service/implement/TodoServiceImplement.java
@@ -18,6 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.OffsetDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -30,18 +31,20 @@ public class TodoServiceImplement implements TodoService {
     @Override
     public ResponseEntity<? super PostTodoResponseDto> saveTodo(PostTodoRequestDto dto, String email) {
         try{
-            if(email == null){
-                return ResponseDto.validationFailed();
-            }
+            if(email == null) return ResponseDto.validationFailed();
+
             MemberEntity member = memberRepository.findByEmail(email);
-            Long memberid = member.getId();
+
+            if(member == null) return PostTodoResponseDto.notExistedUser();
+
+            Long memberId = member.getId();
             OffsetDateTime regDate = OffsetDateTime.now();
             TodoEntity todoEntity = TodoEntity.builder()
                     .title(dto.getTitle())
                     .content(dto.getContent())
                     .state(true)
                     .regDate(regDate)
-                    .memberId(memberid)
+                    .memberId(memberId)
                     .build();
 
             todoRepository.save(todoEntity);
@@ -59,6 +62,8 @@ public class TodoServiceImplement implements TodoService {
             if(email == null) return ResponseDto.validationFailed();
 
             MemberEntity member = memberRepository.findByEmail(email);
+
+            if(member == null) return GetTodoResponseDto.notExistedUser();
             List<TodoEntity> lists = todoRepository.findByMemberId(member.getId());
             todoListItems = lists.stream().map(
                     list-> new TodoListItem(
@@ -81,7 +86,15 @@ public class TodoServiceImplement implements TodoService {
     public ResponseEntity<? super UpdateStateTodoResponseDto> updateState(Long id, UpdateStateTodoRequestDto requestBody, String email) {
         try {
             if(email == null) return ResponseDto.validationFailed();
-            TodoEntity todo = todoRepository.findById(id).orElseThrow();
+
+            MemberEntity member = memberRepository.findByEmail(email);
+
+            if(member == null) return UpdateStateTodoResponseDto.notExistedUser();
+
+            Optional<TodoEntity> optionalTodo = todoRepository.findById(id);
+            if (optionalTodo.isEmpty()) return UpdateStateTodoResponseDto.notExistedTodo();
+
+            TodoEntity todo = optionalTodo.get();
             todo.updateState(requestBody);
         } catch (Exception e){
             e.printStackTrace();
@@ -94,10 +107,16 @@ public class TodoServiceImplement implements TodoService {
     @Transactional
     public ResponseEntity<? super UpdateTodoResponseDto> update(Long id, UpdateTodoRequestDto requestBody, String email) {
         try {
-            if(email == null){
-                return ResponseDto.validationFailed();
-            }
-            TodoEntity todo = todoRepository.findById(id).orElseThrow();
+            if(email == null) return ResponseDto.validationFailed();
+
+            MemberEntity member = memberRepository.findByEmail(email);
+
+            if(member == null) return UpdateTodoResponseDto.notExistedUser();
+
+            Optional<TodoEntity> optionalTodo = todoRepository.findById(id);
+            if (optionalTodo.isEmpty()) return UpdateTodoResponseDto.notExistedTodo();
+
+            TodoEntity todo = optionalTodo.get();
             todo.update(requestBody);
         } catch (Exception e){
             e.printStackTrace();
@@ -110,6 +129,14 @@ public class TodoServiceImplement implements TodoService {
     public ResponseEntity<? super DeleteTodoResponseDto> delete(Long id, String email) {
         try {
             if(email == null) return ResponseDto.validationFailed();
+
+            MemberEntity member = memberRepository.findByEmail(email);
+
+            if(member == null) return DeleteTodoResponseDto.notExistedUser();
+
+            Optional<TodoEntity> optionalTodo = todoRepository.findById(id);
+            if (optionalTodo.isEmpty()) return DeleteTodoResponseDto.notExistedTodo();
+
             todoRepository.deleteById(id);
         }catch (Exception e){
             e.printStackTrace();
